Rework custom SSH key support, load SSH key as required.
This commit reworks the previous custom SSH key support in order to
allow setting the `ansible_ssh_private_key_file` fact. Previously the
customization role allowed setting a SSH *public key* that would be used
for server provisioning. It was assumed that the user would load this
key in their ssh-agent before running Streisand.

This commit updates support to instead proactively set the
`ansible_ssh_private_key_file` for the Streisand SSH key specified. This
ensures Ansible will load & use this private key to access the
streisand-host bypassing the need to use a ssh-agent. Doing so required
changing the customization to prompt for the *private key* and adjusting
other roles accordingly.

Work is done to ensure that the existing-server provisioning SSH test
still functions the way it did when the `streisand` wrapper script was
used to test with a direct `ssh` invocation.