GPG signature verification overhaul. GPG2, HKPS, gpgv2.
This commit revamps the Streisand GPG signature validation and GPG key
management.

Prior to this commit we had a mix of "baked in" GPG keys and
places where the `download-and-verify` playbook was expected to fetch
a key from a keyserver. After this commit a central `gpg` playbook is
used to set up a central keyring with known good keys that can be kept
up to date privately/securely. This is more secure than the code it
replaces, and should be a lot less flaky for end users w.r.t key updates
failing.

Part of this revamp switches to using `gpgv2` to do the actual signature
validation. This is a tool specifically designed for this job and
contains a lot less functionality than the base `gpg2` binary.

In preparation for being able to sign our own content & provide options
for secure communication/filesharing between the Streisand maintainers
and Streisand users this revamp also adds all of the active maintainer's
GPG keys to the Streisand keyring. To ward of any potential suspicion:
The Streisand keyring is separate from the system keyring and can not be
used to verify APT packages and there is no connection between these GPG
public keys and system/ssh access.

The old code had a few things I think were pretty serious deficits:
* It had to rename signature files to expected extensions
* It globbed the inputs to the gpg program in a way that meant one
project's download-and-verify was validating other project's signatures
when they all downloaded to the same dir (e.g. /usr/local/src)
* It used the full-fledged `gpg` binary to validate signatures.
* Some software didn't have a baked in key in the repo so any keyserver
search failures meant a failure to validate a signature
* The key's that were imported from the repo or from keyserver search
were not marked trusted, so all of the signature validations had
warnings about key trust.
* The key ID matching/good signature matching checked the verify output
in a way that could potentially by spoofed by attacker controlled
content in the GPG keys (e.g. the UID strings)
* The default key search timeout was pretty low, and was done over
plaintext HTTP. This seemed to be flaky in practice based on user bug
reports.

New code highlights:

* Uses GPG2 instead of GPG. Modern!
* Builds a standalone keyring for Streisand in `.gnupg/streisand`. Clean!
* Sets up GPG2/dirnmgr to fetch keys over HTTPS with HKPS. Privacy++!
* Increases the key fetch timeouts. Reliability++?
* Sets up a cronjob to refresh the Streisand keyring periodically.
  Up-to-date!
* Pre-initializes keyring with known-good keys from the repo.
  Resiliance++
* Signature verification is done with gpgv2 instead of gpg. Security++
* gpgv2 verification uses exact filenames/paths instead of
  globbing. Security++
* Key ID enforcement is strict, with no chance of matching a key ID from
  a UID field. Security++
* Removed duplicated download/verification code from the Tor playbook.
  Tidy!
* Maybe some other stuff? I forget!