Fix tor-bridge mirror.yml OSX URLs (#984)
It looks like 685a387 was a little bit too thorough and we updated this mirror URL when we shouldn't have. Unfortunately we don't run the tor playbook in CI right now and so the broken Tor Browser Bundle URL didn't surface in CI. This commit restores the correct Tor browser bundle URL for OSX clients.