A script to install Streisand builder dependencies on Ubuntu
One of the stumbling blocks for new users is installing necessary
packages on the builder machine. This script automates it.

Native Python dependencies are installed in `--user` mode, which
will put them under $HOME/.local/bin and lib. Debian users may need to
add ~/.local/bin to their paths.

This script may be suitable for use in cloud-init scripts.