Support specifying custom SSH public key. (#970)
Prior to this PR Streisand universally assumed the user's SSH private key was `~/.ssh/id_rsa`. The public component was then used to provision cloud instances.

This PR adds a new `streisand_ssh_private_key` var to the `global_vars/default-site.yml` file. The site-specific override for these defaults, `~/.streisand/site.yml` is now customized by `playbooks/customize.yml` when the user specifies that they want to use a different SSH key. The public component is expected to exist at the same file path with a `.pub` suffix to keep things simple.

This PR proactively sets the `ansible_ssh_private_key_file` for the Streisand  SSH key specified. This ensures Ansible will load & use this private key to access the streisand-host bypassing the need to use a ssh-agent (though it will play nicely if you do).

The validation playbook is updated to check that the  `streisand_ssh_private_key` exists. The corresponding shell based checks for the SSH key were removed from the `./streisand` wrapper script since they are now replaced by these Ansible tasks. Similarly the existing server playbook
now performs its own quick test of a raw_cmd on the server instead of doing a SSH check in the `./streisand` wrapper. This means things fail a little bit later in the process if you can't actually SSH to the existing server but it isn't *much* later in the process and its nice to `rm` some shell script code.

Two unrelated bugs were fixed with the existing server provisioning:  the genesis role was not being saved as a fact on localhost causing the diagnostics to print "unknown" for the genesis role used. Secondly, because no `gather_facts: yes` for the localhost was run for the existing server provision other diagnostic information was missing. This was also the root cause of Issue #739. 

Resolves #739, #300