Replace id_rsa with just id_rsa.pub and a working ssh-agent
We don't need the private key for anything; what we need is the
capability of logging in. Depend on a working `ssh-agent` that can
unlock `~/id_rsa.pub`.

This PR is not user-friendly now. It requires people who don't
understand ssh to understand `ssh-agent` too. It could be made more
friendly for the old case of an existing `id_rsa`/`id_rsa.pub` by:

1. forcing an `ssh-agent` inside of streisand

2. running `ssh-add` early on, so we can bomb if the user can't unlock
the key.

This is what I do in the Docker builder.

This commit breaks CI, and I don't know how to fix it yet.