Fix wireguard e2e client test.
Previously I was doing way too much work to try and verify the pre and
post wireguard state of `/etc/resolv.conf` on the client. This was
a real pain in the butt and also seems to have broken somewhere down the
road. This commit simplifies the test and ensures it works as expected.
After this commit I can again run a full remote test of an existing
Streisand server provisioned from master without error.