Modular Streisand: improved test coverage of enabled/disabled services (#988)
This PR updates the `tests/tests.sh` driver script so that different build configurations can be run as subtasks depending on a provided env var. The `.travis.yml` configuration is updated to use a matrix of env vars such that we do several subtasks for every CI run. Presently there are 6 build subtasks:

1. A build with everything enabled & all of the syntax checking (the same as was done prior to this commit for all CI runs)
2. A build with just l2tp enabled
3. A build with just openvpn enabled
4. A build with just shadowsocks enabled
5. A build with just the SSH forward user enabled
6. A build with the enabled/disabled services randomized

Care was taken with the randomized build to ensure that there is always 1 service enabled and that none of the dependent services end up in a state that the validation role will reject (e.g. stunnel without openvpn). It is my hope that with the random build we will get some good coverage over time, especially with the daily cron build.

The `development-setup.yml` playbook was updated to only install/build Libreswan on the container host if the container will use l2tp. This should help speed up non-l2tp builds.

Updates #746