Initial OpenVPN end-to-end integration test.
This commit adds an end-to-end integration test for _most_ of the Linux
OpenVPN client profiles. The server common role is updated to write the
randomly picked client names to a machine-readable file to facilitate
this, meaning that a server must have been provisioned with the
testclient var set or the end-to-end openvpn test won't work against it.
Future work could probably address this.

A couple of problematic profiles are not being tested right now. TODO:
Fix these.