Restore sshuttle as optional role with separate user.
This is largely a revert of the portions of aa69d23 that removed
`sshuttle` in order to lock down the ssh forward user `forward`.

Since `sshuttle` doesn't require much from the server-side and to
provide a compromise for users still wanting `sshuttle` this commit
re-introduces `sshuttle` as an optional role that defaults to disabled.

If enabled a separate `sshuttle` user is added to the system and
provided full shell access (but no root).

The validation role is updated to ensure that if sshuttle support is
requested that the ssh_forward support is also enabled. It's possible to
break `sshuttle` out of the `ssh-forward` playbook into its own as
a follow-up since `sshuttle` doesn't rely on the `foward` user anymore.