Avoid hardcoding LE subscriber agreement URL.
Prior to this commit the Let's Encrypt subscriber agreement URL required
for the acmetool non-interactive response file was hardcoded. This means
whenever Let's Encrypt changes their subscriber agreement URL the
lets-encrypt role will break until the agreement var is updated to the
new value.

Fortunately the ACME protocol specifies a way to learn the server's
current terms-of-service agreement URL. Section 7.1.1 of draft-08 says:

    The object MAY additionally contain a field "meta".  If present, it
    MUST be a JSON object; each field in the object is an item of
    metadata relating to the service provided by the ACME server.

    The following metadata items are defined, all of which are OPTIONAL:

      "terms-of-service" (optional, string):  A URL identifying the current
        terms of service.

Let's Encrypt does send a "meta" element with a "terms-of-service" item.
This commit adds a small Python script that fetches the current
terms-of-service URL from the API directory for use in the acmetool
response file.

This will keep the response file current even if the agreement URL
changes.