Gateway: generate a custom `openssl.cnf` to make certs taste better
Windows is pretty unhappy with the certs we were generating. Both
Chrome and Edge on Windows 10 would not even let you navigate to the
page, and don't present an "advanced / I'm sure" option.

This is a WIP. It needs a cleanup of `openssl.cnf`; all the working parts
are there, but there's a bunch of stuff we don't use.

Note to self: Remember to credit
https://github.com/stanzgy/wiki/blob/master/network/openssl-self-signed-certs-cheatsheet.md

commit 6742820570dca5575e0da2881326f86c7d63d1b9
Author: Jay Carlson <nop@nop.com>
Date:   Tue Nov 21 11:35:33 2017 -0500

    Break up OpenSSL commands into multiple lines