Ensure dnsmasq is reloaded for WG.
In some configurations (Only WireGuard enabled for instance) it seems
that the `notify: Restart dnsmasq` that the WireGuard role was using to
ensure its `/etc/dnsmasq.d/` config file was being loaded was skipped by
Ansible. If nothing else restarts `dnsmasq` then the WireGuard client DNS
is broken because `dnsmasq` doesn't load the WireGuard config fragment
and listen on the wg0-server interface.

I can't seem to understand _why_ Ansible thinks it can skip this notify
so as a short-term fix this commit adds an explicit `service` task to
the WireGuard role that reloads `dnsmasq` after the config fragment file
is written.