Direct link to NetworkManager OpenVPN files.
Previous to this commit the "Linux (Ubuntu)" NetworkManager instructions
for configuring OpenVPN with a client profile instructed users to
download several files from a raw directory index. This broke in 811be
when we added `autoindex off;` to the nginx config.

This commit updates the OpenVPN instructions template (for EN and FR) to
directly link to each of the required files instead of relying on the
directory index. This is a better user experience overall, directory
indexes are ugly!

My FR updates are probably très mauvais. Je suis désolé!! :-x