Test tor relay w/ obfs4proxy & gateway hidden service.
This commit adds an end-to-end integration test for the tor services of
a Streisand server. Using the Android QR code from the gateway the test
bootstraps a torrc config that uses the remote Streisand server as
a bridge relay w/ obfs4 obfuscation. The test also checks that the
Streisand gateway's hidden service is up & enforcing HTTP basic
auth, and can be visited successfully using the correct
username/password. In order to test the hidden service the Streisand
server must be generated with `streisand_client_test: yes` so that the
hidden service address is written somewhere in the HTTPS webroot for the
test client to download programmatically.