virtualenv requirements fix (#1102)
Re-order the requirements for virtualenv pip requirements so that we get the correct dependencies for Azure and Linode.

Since `pip install ansibile[azure]` installs the latest stable ansible as well as the necessary azure requirements, explicitly requiring `ansible` is no longer necessary; doing so will cause errors when deploying to Azure.

Additionally, the `linode-python` module makes no mention of requiring the `pycurl` module and therefore does not get installed. As such, `pycurl` is explicitly declared.