End-to-end integration tests for WireGuard, OpenVPN, OpenConnect, Tor, Stunnel (#1047)
## WireGuard
The new improved WireGuard integration test now properly tests that a successful handshake between the client and server occurs. It will also check that the WireGuard DNS is added and removed from `/etc/resolv.conf` successfully. The integration test checks each of the generated WireGuard profiles to ensure each works correctly.

## OpenVPN
The new OpenVPN integration test checks all of the randomly generated client profiles across the different OpenVPN configurations we offer. Similar to the Wireguard tests it will check that a successful connection can be made, that the DNS is updated as expected, and that a remote website is accessible with the VPN up. At present only the `direct` and `sslh` type profiles are tested. The `direct-udp` and `combined` profiles aren't tested - they weren't working for me and I haven't yet determined if it's due to my peculiar home network or if there are bugs at work.

## OpenConnect
The new OpenConnect integration test checks all of the randomly generated PKCS12 client profiles. It learns the PKCS12 passwords from the AnyConnect `.mobileconfig` XML files. Similar to the other services we test that DNS is modified as expected & that an Internet resource is accessible over the service.

## Tor
The Tor integration test checks that the obfs4proxy relay on a Streisand service works correctly as a bridge relay for the test client. We also test that the Streisand hidden service is accessible, enforcing basic auth, and serving the homepage to users that know the username/password.

## Stunnel
The Stunnel integration test checks that the stunnel4 client configuration works as advertised & that the OpenVPN profiles that target the stunnel4 proxy work too. This test imports the OpenVPN profile testing mechanics to test that the DNS is updated & the overall Stunnel wrapped VPN connection works.