End-to-end integration test for TinyProxy. (#1112)
This commit adds an end-to-end integration test for TinyProxy. In the
process the SSH forward user integration test was rewritten from being
a bash script to using Ansible. The TinyProxy test is added to the end
of the SSH forward integration test because accessing TinyProxy depends
on using the SSH forward user to port forward the TinyProxy port.