Remove deprecated LXD PPA, use snap to install LXD. (#1156)
The upstream LXD project has deprecated the PPA we were using to get
a modern LXD/LXC version installed in Travis' super crusty 14.04
instances. The official recommendation is to either: A) use the old
packaged version in 14.04 or B) install a new version using `snap`.

Since we rely on network features not available in the 14.04 packaged
LXD this commit adopts the Snap approach (thanks to @alimakki for
originally starting on this!).

The primary changes involve configuring everything with a non-standard
LXD UNIX domain socket location.