Gateway: generate a custom `openssl.cnf` to make certs taste better (#1063)
Windows is pretty unhappy with the certs we were generating. Both
Chrome and Edge on Windows 10 would not even let you navigate to the
page, and don't present an "advanced / I'm sure" option. Much of this is
due to the prior openssl config. This commit cleans up the `openssl.cnf`
trying to maximize compatibility with browsers.