Use distinct name for stunnel GCP fw rule. (#1183)
Prior to this commit the GCP stunnel firewall rule was created under the
same name as the OpenVPN firewall rule. This would stomp the OpenVPN
ports when stunnel was enabled. This commit uses a distinct name for the
stunnel firewall rule so that both it and the base OpenVPN rule can
coexist.