Update `syntax-check.yml`.
There's not a convenient way to dynamically import a playbook so
unfortunately we're left maintaining this by hand. This commit updates
the playbooks and roles that were not being syntax checked. This proved
to be worthwhile because it found one real bug in the pkcs tasks of the
certificates role!