Update `syntax-check.yml` (#1187)
There's not a convenient way to dynamically import a playbook so
unfortunately we're left maintaining `syntax-check.yml` by hand. This commit
updates the playbooks and roles that were not being syntax checked. This 
proved to be worthwhile because it found one real bug in the `pkcs` tasks of the
`certificates` role!

Resolves https://github.com/StreisandEffect/streisand/issues/1184