Convert SSLH to use config file vs cmd line params. (#1219)
This commit updates the `sslh` playbook such that the
`/etc/default/sslh` systemd defaults specify a `-F` flag with
a configuration file instead of a bunch of individual command line
options. The new configuration is read from `/etc/sslh.cfg` and contains
the same options as were previously specified by command line.

This is slightly cleaner overall and will be easier to extend in the
future. This also unblocks the usage of `log_level` when SSLH is
upgraded to a version > 1.18.

Note: specifying `probe: "builtin"` for each protocol is only 
required [pre 1.18](https://github.com/yrutschle/sslh/issues/121#issuecomment-341899139). For the version shipped in Streisand currently (1.17-2) 
it is required.

Resolves https://github.com/StreisandEffect/streisand/issues/827