shadowsocks: install shadowsocks-libev from PPA.
This commit replaces the Ansible code that built `shadowsocks-libev`
from source with code to add an upstream PPA. This will ensure that
security updates are installed automatically and will speed up server
provisioning.

The build dependencies are still installed because `simple-obfs` is
being built from source (it is not presently available in an upstream
PPA). It's likely that this is installing more packages than are
required to build only `simple-obfs` but my hope is in the near future
it will be replaced by PPA install as well.