GPG signature verification overhaul. GPG2, HKPS, gpgv2. (#1231)
This commit revamps the Streisand GPG signature validation and GPG key management.

Prior to this commit we had a mix of "baked in" GPG keys and places where the `download-and-verify` playbook was expected to fetch a key from a keyserver. After this commit a central `gpg` playbook is used to set up a keyring with known good keys that can be kept up to date. Overall this approach should be more secure and hopefully a lot less flaky for end users experiencing flaky keyserver errors.

Part of this revamp switches to using `gpgv2` to do the actual signature validation. This is a tool [specifically designed for this job](https://www.gnupg.org/documentation/manuals/gnupg/gpgv.html) and contains a lot less functionality than the base `gpg2` binary. One of the handy things about `gpgv2` is it assumes the keys in the provided keyring are ultimately trusted. In contrast, when using `gpg2` to verify signatures you have to either do the whole `gpg2 edit <keyid>`, `trust`, `5`, `save` dance for each key (interactive, annoying), or do that once and then use `--export-ownertrust` and `--import-ownertrust` (workable, but more code!).

In preparation for being able to sign our own content & provide options for secure communication/filesharing between the Streisand maintainers and Streisand users this revamp also adds all of the active maintainer's GPG keys to the Streisand keyring. _To ward off any potential suspicion: The Streisand keyring is separate from the system keyring and can not be used to verify APT packages and there is no connection between these GPG public keys and system/ssh access._

The old code had a few things I think were downsides:
* It had to rename signature files to expected extensions
* It globbed the inputs to the `gpg` binary in a way that meant one project's `download-and-verify` run was validating other project's signatures when they all downloaded to the same dir (e.g. `/usr/local/src`)
* It used the full-fledged `gpg` binary to validate signatures.
* Some software didn't have a baked in key in the repo so any keyserver search failures meant a failure to validate a signature and a provisioning error
* The key's that were imported from the repo or from keyserver search were not marked trusted, so all of the signature validations had warnings about key trust.
* The key ID matching/good signature matching checked the verify output in a way that could potentially by spoofed by attacker controlled content in the GPG keys (e.g. the UID strings)
* The default key search timeout was pretty low, and key search was done over plaintext HTTP on a non-standard port. This seemed to be flaky in practice based on user bug reports.

New code highlights:

* Uses GPG2 instead of GPG. Modern!
* Builds a standalone keyring for Streisand in `.gnupg/streisand`. Clean!
* Sets up GPG2/`dirmngr` to fetch keys over HTTPS with HKPS to the [HKPS Keyservers.net pool](https://sks-keyservers.net/overview-of-pools.php#pool_hkps). Privacy++!
* Increases the key fetch timeouts. Reliability++?
* Sets up a cronjob to refresh the Streisand keyring periodically. Up-to-date!
* Pre-initializes keyring with known-good keys from the repo. Resilience++
* Signature verification is done with `gpgv2` instead of `gpg`. Security++
* `gpgv2` verification uses exact filenames/paths instead of globbing. Security++
* Key ID enforcement is strict, with no chance of matching a key ID from a UID field. Security++
* Duplicated download/verification code from the `tor-bridge` playbook was replaced by `download-and-verify` usage. Tidy!
* Maybe some other stuff? I forget!

The following GPG keys are imported to the Streisand keyring:
```
# OpenVPN release signing key
- 2F2B01E7.security@openvpn.net.asc
# Openconnect
- 7F343FA7.nmav@redhat.com.asc
- 96865171.nmav@gnutls.org.asc
# Tor browser release signing key
- 93298290.torbrowser@torproject.org.asc
# PuTTY release signing key
- B43434E4.putty@projects.tartarus.org.asc
# Stunnel release signing key
- DD3AAAA3.Michal.Trojnara@stunnel.org.asc
# LibreSwan release signing key
- B30FC6F9.team@libreswan.org.asc
# Streisand maintainer - Github @cpu
- 2D8330C2.daniel@binaryparadox.net.asc
# Streisand maintainer - Github @jlund
- CDF6583E.josh@joshlund.com.asc
# Streisand maintainer - Github @nopdotcom
- F67DA905.nop@nop.com.asc
# Streisand maintainer - Github @CorbanR
- 3F7F585B.corban@raunco.co.asc
```

I've left out @alimakki's key temporarily because it seems to mysteriously break in CI but works when testing real hosts :confused: :question: 

For each piece of respective software we pin the specific key ID from the keyring for verifying releases:
* OpenVPN -> `AF131CAE`
* OpenConnect -> `9013B842`
* PuTTY -> `B43434E4`
* Stunnel -> `D416E014`
* Libreswan -> `B30FC6F9`

It's safe to refer to these keys by short ID prefixes strictly because we are using a prepopulated and separated GPG keyring and not searching keyservers for the initial import.

Resolves #1199, #1227, #963