Temporary pins of Ansible, SecretStorage versions
Ansible 2.5.1 appears to be broken for us. 2.5.2 looks like it will
have a fix. Pin to version 2.5.0 until then.

The SecretStorage module was just bumped to 3.0.0, and now requires
Python 3. Until we're ready for Python 3, specify the earlier rev.

GCE may require pycrypto for authentication now. Leave a note about
that; we'll uncomment after another round of testing.