GPG: explicitly wait for dirmngr and gpg-agent sockets.
A user provisioning Streisand on Azure reports a persistent failure
during the Streisand GPG keyring import task with the following error
msg:
  "gpg: can't connect to the agent: IPC connect call failed"
This is (likely?) caused by the gpg-agent not having started by the time
we import the keys.

This commit changes the gpg playbook to explicitly start both the
gpg-agent and the dirnmgr instead of letting the first gpg operation do
so. This commit also adds an explicit `wait_for` task for both the
dirnmgr and gpg agent control sockets to make sure we don't proceed
until both daemons are available.