GPG: Use explicit English locale for gpgv2.
The download-and-verify playbook invokes `gpg2` to check GPG signatures.
In addition to checking the return code for success we regex match the
key ID that was used to successfully validate the signature. The
successful match of this regex depends on the `gpgv2` output being in
English. If a user has override their `LANG` to something else then the
regex will fail to match! See:
https://github.com/StreisandEffect/streisand/issues/1199#issuecomment-378677563

This commit addresses this case by explicitly forcing the locale to
"en_US.UTF-8" when invoking `gpgv2`, ignoring the system locale
configuration.

Its likely bugs like this exist other places in Streisand but at least
this one is fixed!