GPG: explicitly wait for dirmngr and gpg-agent sockets. (#1314)
* GPG: explicitly wait for dirmngr and gpg-agent sockets.

A user provisioning Streisand on Azure reports a persistent failure
during the Streisand GPG keyring import task with the following error
msg:
  "gpg: can't connect to the agent: IPC connect call failed"
This is (likely?) caused by the gpg-agent not having started by the time
we import the keys.

This commit changes the gpg playbook to explicitly start both the
gpg-agent and the dirnmgr instead of letting the first gpg operation do
so. This commit also adds an explicit `wait_for` task for both the
dirnmgr and gpg agent control sockets to make sure we don't proceed
until both daemons are available.

* GPG: Use explicit English locale for gpgv2.

The download-and-verify playbook invokes `gpg2` to check GPG signatures.
In addition to checking the return code for success we regex match the
key ID that was used to successfully validate the signature. The
successful match of this regex depends on the `gpgv2` output being in
English. If a user has override their `LANG` to something else then the
regex will fail to match! See:
https://github.com/StreisandEffect/streisand/issues/1199#issuecomment-378677563

This commit addresses this case by explicitly forcing the locale to
"en_US.UTF-8" when invoking `gpgv2`, ignoring the system locale
configuration.

Its likely bugs like this exist other places in Streisand but at least
this one is fixed!