wireguard: remove awkward client/server notation
This had with it a few problems. Some versions of android don't like
config files with a dot in them, it would increase the length of the
interface beyond what it should be, and generally this belies the real
architecture of wireguard. Instead, call test interfaces and the server
interfaces 'wg0' and call each client interface by its profile name
alone.

While we're at it, we clean up the instructions a bit.