While asking for LE FQDN, tell the user what address it should point at
If you let Streisand create a server for you, it will not have a FQDN
pointing at it. For new Ansible users, it isn't immediately obvious
where to find the IP address of the new server. Be explicit.

If there is a better way to get the IP address at this stage, please fix.