Ensure VPN profile names fit in Linux network device names (#1400)
There's a 15-character limit on the length of Linux network device names, and it would be nice to use the VPN profile names as interface names. A few of the wordlist words are 8 letters long. We can use two words as long as they're each 7 characters or shorter (leaving room for the separator).