Add a 10.10.10.10 service network (#1462)
It would be nice to have a place all authenticated users have access
to, particularly for DNS right now. One inspiration is that the
systemd resolver lives at 127.0.0.53. But we can't tell VPN clients to
do DNS at 127.0.0.53 because of course that's their own loopback
interface.

This sets up a dummy network 10.10.10.0/24, and sets up the service0
10.10.10.10 interface. dnsmasq listens there. Non-53 DNS service can
live there too.

The logical next step is to change all the advertisesd DNS addresses
to 10.10.10.10. This would require a round of testing of all those
protocols.

Simplifying DNS configuration may make it easier to replace or
supplement the service provided (or at least makes look
easier). Installing the `unbound` resolver on 10.10.10.10 sounds easy,
although it's probably better to have dnsmasq delegate to unbound.

Fascist server owners could force *all* port 53 traffic to
10.10.10.10.