Run tor HS as separate tor process/user, make HS use local bridge
The Tor Project recommends against running a hidden service and
a tor bridge on the same machine, or at the very least to run them
as separate processes. This way, the HS activity won't be able to
block the bridge's activity (this is a side effect of Tor being
single-threaded). Also, the bridge will actually report your
onion service history in its statistics if they use the same process.

See:

https://trac.torproject.org/projects/tor/ticket/8742
https://trac.torproject.org/projects/tor/ticket/16585

for more info.

Also, since Streisand is making a locally running bridge, it makes
sense for the HS to use that same bridge to connect to Tor.