ad-blocking via dnsmasq; disable cloudflared in CI
This PR implements DNS-level blocking of ads and trackers. Dnsmasq is configured to return `0.0.0.0` for bad domains.

notracking/hosts-blocklists maintains easy-to-use network block lists. Importantly, it supports both host and domain blacklisting. The lists are refreshed daily.

To reduce the mischief that upstream could cause, files downloaded are rewritten through awk filters. Notably, the domain blocker part ensures there are no directives in the dnsmasq config fragment other than address.

As a reminder: Streisand DNS services are available to authenticated users at `10.10.10.10` (among other addresses).

Cloudflared is now disabled in CI. I'm not thrilled with this, but runs with the `cloudflared` target simply weren't completing. A PR turning it back on is welcome.